requirements.txt
Flask
azure-storage-blob
opencv-python
face_recognition